requirements.txt
# flask essentials
flask
flask-login
flask-security-too[fsqla, mfa]

# MySQL and a dependent library
pymysql
cryptography


# Password hashing
argon2_cffi
# Password Strength Checker
zxcvbn

# Restful API
Flask-RESTful

# Test
pytest
# API Test Framework
tavern

# Guideline Enforcement
pylint
flake8
mccabe


# SAST
bandit
# pyre

## pytest-cov
# flask_mysqldb
# flask-wtf
# flask_sqlalchemy